Add v1 maplibre-gl implementation of map under flag
it's missing all custom layers at this point
