+++ frontend/package.json
@@ -32,6 +32,7 @@
     "immer": "^9.0.16",
     "lodash": "^4.17.21",
     "mapbox-gl": "^2.4.1",
+    "maplibre-gl": "^5.6.2",
     "netlify-identity-widget": "^1.9.2",
     "normalize.css": "^8.0.1",
     "qs": "^6.11.0",
@@ -69,6 +70,7 @@
     "@types/lint-staged": "~13.3.0",
     "@types/lodash": "^4.17.13",
     "@types/mapbox-gl": "^1.13.0",
+    "@types/maplibre-gl": "^1.13.2",
     "@types/mini-css-extract-plugin": "^2.5.1",
     "@types/netlify-identity-widget": "^1.9.3",
     "@types/prettier": "^2.7.1",
