+++ frontend/src/screens/App/screens/MapPane/components/MainMap/index.tsx
@@ -1,29 +1,18 @@
 import React from 'react';
+import { useFeatureFlag } from 'screens/App/shared/stores/FeatureFlagsStore';
+import FeatureFlag from 'screens/App/shared/types/FeatureFlag';
 
-import {
-  // withRouter,
-  useHistory,
-  useLocation,
-  useRouteMatch,
-} from 'react-router-dom';
-
-import classnames from 'classnames';
-import mapboxgl from 'mapbox-gl';
+// Import both implementations
+import MapBoxMap from './MapBoxMap';
+import MapLibreMap from './MapLibreMap';
 
 import LoginForm from 'shared/components/LoginForm';
 import useLoginStore from 'shared/stores/LoginStore';
+import { MapInterface } from './MapInterface';
 import * as overlays from './overlays';
 
+// Export the OverlayId type from the appropriate module
 export { OverlayId } from './overlays';
-
-import { RouteComponentProps } from 'react-router';
-import stylesheet from './MainMap.less';
-
-const MAPBOX_STYLE = __DEV__
-  ? 'mapbox://styles/julianboilen/ck5jrzrs11r1p1imia7qzjkm1/draft'
-  : 'mapbox://styles/julianboilen/ck5jrzrs11r1p1imia7qzjkm1';
-
-const PHOTO_LAYER = 'photos-1940s';
 
 interface Props {
   className?: string;
@@ -32,179 +21,50 @@ interface Props {
   requireLogin?: boolean;
 }
 
-type PropsWithRouter = Props & RouteComponentProps<{ identifier?: string }>;
+// Simple component that switches between implementations
+const MainMap = React.forwardRef<MapInterface, Props>((props, ref) => {
+  const useMapLibre = useFeatureFlag(FeatureFlag.USE_MAPLIBRE);
+  const { isLoggedInToNonAnonymousAccount, isLoadingMe } = useLoginStore();
 
-class MainMap extends React.PureComponent<PropsWithRouter> {
-  private mapContainer: HTMLElement;
-  private map: mapboxgl.Map;
+  // Only show map if user is logged in to a non-anonymous account
+  // But allow override if requireLogin is explicitly set to false
+  const canShowMap =
+    props.requireLogin === false ||
+    (isLoggedInToNonAnonymousAccount && !isLoadingMe);
 
-  componentDidMount(): void {
-    const map = (this.map = new mapboxgl.Map({
-      container: this.mapContainer,
-      style: MAPBOX_STYLE,
-      center: [-73.99397, 40.7093],
-      zoom: 13.69,
-      maxBounds: [
-        [-74.25908989999999, 40.4773991], // SW
-        [-73.70027209999999, 40.9175771], // NE
-      ],
-      hash: true,
-    }));
-
-    map.on('click', PHOTO_LAYER, (e) => {
-      const { panOnClick } = this.props;
-      if (panOnClick) map.panTo(e.lngLat);
-      if (!e || !e.features) return;
-      const feature = e.features[0];
-      const identifier = feature.properties?.photoIdentifier as string;
-      this.props.history.push({
-        pathname: '/map/photo/' + identifier,
-        hash: window.location.hash,
-      });
-    });
-
-    // Change the cursor to a pointer when the mouse is over the places layer.
-    map.on('mouseenter', PHOTO_LAYER, () => {
-      map.getCanvas().style.cursor = 'pointer';
-    });
-
-    // Change it back to a pointer when it leaves.
-    map.on('mouseleave', PHOTO_LAYER, () => {
-      map.getCanvas().style.cursor = '';
-    });
-
-    map.on('style.load', () => {
-      overlays.installLayers(this.map, PHOTO_LAYER);
-
-      map.setLayoutProperty(PHOTO_LAYER + '-active', 'visibility', 'visible');
-
-      this.syncUI();
-    });
-
-    // Added to remove layers outside the viewport, to make the attribution correct
-    map.on('moveend', () => {
-      if (!map.isStyleLoaded()) return;
-      this.syncUI();
-    });
-  }
-
-  componentDidUpdate(prevProps: PropsWithRouter): void {
-    // Update the conditional color expression to make the active dot a different color
-    if (!this.map.isStyleLoaded()) {
-      this.map.once('style.load', () => this.syncUI());
-    }
-    if (
-      prevProps.match.params.identifier !==
-        this.props.match.params.identifier ||
-      prevProps.overlay !== this.props.overlay
-    ) {
-      this.syncUI();
-    }
-  }
-
-  syncUI(): void {
-    this.map.setFilter(PHOTO_LAYER + '-active', [
-      '==',
-      ['get', 'photoIdentifier'],
-      this.props.match.params.identifier || null,
-    ]);
-
-    overlays.setOverlay(this.map, this.props.overlay);
-  }
-
-  componentWillUnmount(): void {
-    if (this.map) this.map.remove();
-  }
-
-  /**
-   * Call if container has resized
-   */
-  resize(): void {
-    this.map.resize();
-  }
-
-  goTo(center: mapboxgl.LngLatLike): void {
-    this.map.easeTo({
-      zoom: 17.5,
-      center,
-    });
-  }
-
-  render(): React.ReactNode {
-    const { className: propsClassName } = this.props;
-
+  if (!canShowMap) {
     return (
       <div
-        className={classnames(stylesheet.map, propsClassName)}
-        ref={(el) => (this.mapContainer = el)}
-      />
-    );
-  }
-}
-
-/* This component needs to have public methods. Regular withRouter does not forward the ref needed to use public methods.
-Created my own withRouter that forwards refs.
-All the types got very confusing, sorry.
-*/
-
-// eslint-disable-next-line @typescript-eslint/explicit-function-return-type
-function withRouterRef<
-  OuterProps extends Props,
-  C = React.ComponentType<OuterProps & RouteComponentProps>
->(Component: C) {
-  return React.forwardRef<C, OuterProps>(function WithRouterRef(props, ref) {
-    const match = useRouteMatch();
-    const location = useLocation();
-    const history = useHistory();
-    const { isLoggedInToNonAnonymousAccount, isLoadingMe } = useLoginStore();
-
-    // Only show map if user is logged in to a non-anonymous account
-    // But allow override if requireLogin is explicitly set to false
-    const canShowMap =
-      props.requireLogin === false ||
-      (isLoggedInToNonAnonymousAccount && !isLoadingMe);
-
-    if (!canShowMap) {
-      return (
+        style={{
+          height: '100%',
+          display: 'flex',
+          alignItems: 'center',
+          justifyContent: 'center',
+          backgroundColor: '#f5f5f5',
+          padding: '2rem',
+        }}
+      >
         <div
           style={{
-            height: '100%',
-            display: 'flex',
-            alignItems: 'center',
-            justifyContent: 'center',
-            backgroundColor: '#f5f5f5',
-            padding: '2rem',
+            textAlign: 'center',
+            maxWidth: '400px',
           }}
         >
-          <div
-            style={{
-              textAlign: 'center',
-              maxWidth: '400px',
-            }}
-          >
-            <h2>Login Required</h2>
-            <p>Please log in to view the map.</p>
-            <LoginForm newEmailBehavior="create" />
-          </div>
+          <h2>Login Required</h2>
+          <p>Please log in to view the map.</p>
+          <LoginForm newEmailBehavior="create" />
         </div>
-      );
-    }
-
-    return (
-      // eslint-disable-next-line @typescript-eslint/ban-ts-comment -- I give up on types
-      // @ts-ignore
-      <Component
-        ref={ref}
-        match={match}
-        location={location}
-        history={history}
-        {...props}
-      />
+      </div>
     );
-  });
-}
+  }
 
-export default withRouterRef<Props>(MainMap) as unknown as React.ComponentType<
-  Props & React.RefAttributes<MainMap>
-> &
-  Pick<MainMap, 'goTo' | 'resize'>;
+  if (useMapLibre) {
+    return <MapLibreMap ref={ref} {...props} />;
+  } else {
+    return <MapBoxMap ref={ref} {...props} />;
+  }
+});
+
+MainMap.displayName = 'MainMap';
+
+export default MainMap;
