+++ frontend/src/screens/App/shared/types/FeatureFlag.ts
@@ -2,6 +2,7 @@ enum FeatureFlag {
   AUDIO_STORYTELLING = 'audio-storytelling',
   DISABLE_WELCOME_MODAL = 'disable-welcome-modal',
   ALWAYS_SHOW_OVERLAY = 'always-show-overlay',
+  USE_MAPLIBRE = 'use-maplibre',
 }
 
 export default FeatureFlag;
